1/25 dotnet 修复部分由于数据库迁移产生的BUG +1

+++ dotNetCore/Dal/SqlHelper.cs
@@ -94,7 +94,7 @@ namespace Dal
                 {
 
                     cmd.CommandType = CommandType.Text;
-                    res = (int)cmd.ExecuteScalar();
+                    res = (int)(long)cmd.ExecuteScalar();
                 }
             }
             catch (Exception ex)
@@ -130,7 +130,7 @@ namespace Dal
                 {
 
                     cmd.CommandType = CommandType.Text;
-                    res = (int)cmd.ExecuteScalar();
+                    res = (int)(long)cmd.ExecuteScalar();
                 }
             }
             catch (Exception ex)
@@ -290,7 +290,7 @@ namespace Dal
                 {
                     cmd.CommandType = CommandType.Text;
                     cmd.Parameters.Add(new MySqlParameter("@number", number));
-                    res = (int)cmd.ExecuteScalar();
+                    res = (int)(long)cmd.ExecuteScalar();
                 }
             }
             catch (Exception ex)
